refactor: move search result to search module
