+++ src/main/java/com/find/doongji/apt/client/AptClient.java
@@ -1,7 +1,7 @@
 package com.find.doongji.apt.client;
 
 import com.find.doongji.apt.payload.response.DanjiCode;
-import com.find.doongji.apt.payload.response.SearchResult;
+import com.find.doongji.search.payload.response.SearchResult;
 
 import java.util.List;
