+++ src/main/java/com/find/doongji/apt/client/AptDetailClient.java
@@ -1,7 +1,7 @@
 package com.find.doongji.apt.client;
 
 import com.find.doongji.apt.payload.response.DanjiCode;
-import com.find.doongji.apt.payload.response.SearchResult;
+import com.find.doongji.search.payload.response.SearchResult;
 import com.find.doongji.apt.utils.HttpUtils;
 import com.find.doongji.apt.utils.ParseUtils;
 import org.springframework.beans.factory.annotation.Value;
