--- src/main/java/com/find/doongji/recommend/RecommendController.java
@@ -1,37 +1,0 @@
-package com.find.doongji.recommend;
-
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.client.RestTemplate;
-
-@RestController
-@RequestMapping("/api/v1/recommend")
-public class RecommendController {
-
-    @PostMapping("/search")
-    public ResponseEntity<?> findSimilar(@RequestParam("query") String query, @RequestParam(value = "top_k", defaultValue = "16384") int topK) {
-        try {
-            String fastApiUrl = "http://localhost:8001/find_similar";
-
-            HttpHeaders headers = new HttpHeaders();
-            headers.setContentType(MediaType.APPLICATION_JSON);
-
-            String requestBody = String.format("{\"query\": \"%s\", \"top_k\": %d}", query, topK);
-
-            HttpEntity<String> requestEntity = new HttpEntity<>(requestBody, headers);
-
-            RestTemplate restTemplate = new RestTemplate();
-            ResponseEntity<String> response = restTemplate.postForEntity(fastApiUrl, requestEntity, String.class);
-
-            return ResponseEntity.ok(response.getBody());
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error: " + e.getMessage());
-        }
-    }
-}
