+++ src/main/java/com/find/doongji/search/service/SearchService.java
@@ -1,6 +1,6 @@
 package com.find.doongji.search.service;
 
-import com.find.doongji.apt.payload.response.SearchResult;
+import com.find.doongji.search.payload.response.SearchResult;
 import com.find.doongji.search.payload.request.SearchRequest;
 
 import java.util.List;
